United in Duty, Bound in Destiny.
On Mata Nui, I learned that no one stood alone.
On Metru Nui, I learned that actions, not fate, defined a hero.
On Voya Nui, I learned what it means to be determined, to face your fears.
On Mahri Nui, I learned what sacrifice truly means.
On Karda Nui, I learned to never give up, even in the Darkest Hour.
On Bara Magna, I learned that anyone, anywhere, can be a hero.

United in Duty, Bound in Destiny.

That is the way of the Bionicle.